Update boot.py - Add auto-boundary
I think it might be cheap insurance to have it on for all presets, as it certainly can help some datasets that would otherwise fail to reconstruct to reconstruct properly. It has little benefit for at-nadir surveys and processing templates, but sometimes these datasets can have unexpectedtly large reconstruction areas above/below the plane of reconstruction due to things like water or other spectral material in the survey, and this should prevent that edge case from being fatal.