Add files via upload
This is a notebook that allows the user to iterate ground classifier parameters and visualize results for optimizing the algorithm of their choice using PDAL. Caveat- this notebook requires hard paths and the knowledge to build a Python environment and import packages. This notebook is also designed for SfM RGB point clouds. I will be adding another notebook for visualizing Micasense multispec point clouds.